@@ -64,9 +64,9 @@ class MyApp extends StatelessWidget {
       theme: themeData,
       builder: (context, child) {
         return StreamChat(
-          child: child,
           client: client,
           streamChatThemeData: customTheme,
+          child: child,
         );
       },
       home: ChannelListPage(),
@@ -82,7 +82,7 @@ class ChannelListPage extends StatelessWidget {
         child: ChannelListView(
           filter: Filter.in_(
             'members',
-            [StreamChat.of(context).user.id],
+            [StreamChat.of(context).user!.id],
           ),
           sort: [SortOption('last_message_at')],
           pagination: PaginationParams(
@@ -97,7 +97,7 @@ class ChannelListPage extends StatelessWidget {
 
 class ChannelPage extends StatelessWidget {
   const ChannelPage({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -123,10 +123,10 @@ class ChannelPage extends StatelessWidget {
 }
 
 class ThreadPage extends StatelessWidget {
-  final Message parent;
+  final Message? parent;
 
   ThreadPage({
-    Key key,
+    Key? key,
     this.parent,
   }) : super(key: key);
 
@@ -134,7 +134,7 @@ class ThreadPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: ThreadHeader(
-        parent: parent,
+        parent: parent!,
       ),
       body: Column(
         children: <Widget>[
